添加Activiti导出图片接口，优化Activiti创建任务及完成任务的接口

+++ service-feign/src/main/java/com/yfny/servicefeign/fallback/ExampleHelloServiceHystric.java
@@ -38,7 +38,7 @@ public class ExampleHelloServiceHystric implements ExampleHelloService {
     }
 
     @Override
-    public int auditDemand(int demandId, String taskId, String auditOpinion, String shrId, String orgId, boolean pass) {
+    public int auditDemand(Long demandId, String taskId, String auditOpinion, String shrId, String orgId, boolean pass) {
         return 0;
     }
